add a helper function to alloc SSE/AVX ps blocks by default
