+++ src/common/sse.h
@@ -16,6 +16,26 @@
 
 #include <xmmintrin.h>
 
+#include "common/darktable.h"
+
+__DT_CLONE_TARGETS__
+static inline void dt_simd_memcpy(const float *const restrict in,
+                                  float *const restrict out,
+                                  const size_t num_elem)
+{
+  // Perform a parallel vectorized memcpy on 64-bits aligned
+  // contiguous buffers. This is several times faster than the original memcpy
+
+#ifdef _OPENMP
+#pragma omp parallel for simd default(none) \
+dt_omp_firstprivate(in, out, num_elem) \
+schedule(static) aligned(in, out:64)
+#endif
+  for(size_t k = 0; k < num_elem; k++)
+    out[k] = in[k];
+}
+
+
 /**
  * Fast SSE2 implementation of special math functions.
  */
